+++ packages/svelte/src/div/index.ts
@@ -1,0 +1,1 @@
+export { default as EventStopper } from "./EventStopper.svelte";