+++ packages/svelte/src/index.ts
@@ -1,5 +1,7 @@
 export * from "./image";
 export * from "./video";
 export * from "./lazy";
+// export * from "./button";
+export * from "./div";
 
 export const Name = "@a2-ts-utils/svelte";